server/requirements.txt
# Minimal server dependencies (Prompt 1 scaffold)
# TODO: Pin versions appropriately and review optional deps.
fastapi
uvicorn
sqlmodel
httpx
python-dotenv
pytest
pytest-cov

# LangChain split packages
langchain-core
langchain

# Providers (optional, installed as needed)
langchain-openai
langchain-anthropic
langchain-ollama
langchain-community
